+++ allennlp/data/dataset_readers/__init__.py
@@ -8,7 +8,11 @@ and produce datasets in the formats required by specific models.
 
 
 from allennlp.data.dataset_readers.conll2003 import Conll2003DatasetReader
-from allennlp.data.dataset_readers.dataset_reader import DatasetReader
+from allennlp.data.dataset_readers.dataset_reader import (
+    DatasetReader,
+    AllennlpDataset,
+    AllennlpLazyDataset,
+)
 from allennlp.data.dataset_readers.interleaving_dataset_reader import InterleavingDatasetReader
 from allennlp.data.dataset_readers.sequence_tagging import SequenceTaggingDatasetReader
 from allennlp.data.dataset_readers.sharded_dataset_reader import ShardedDatasetReader