+++ tests/common/file_utils_test.py
@@ -16,6 +16,7 @@ from allennlp.common.file_utils import (
     cached_path,
     _split_s3_path,
     open_compressed,
+    CacheFile,
 )
 from allennlp.common.testing import AllenNlpTestCase
 
@@ -251,3 +252,13 @@ class TestFileUtils(AllenNlpTestCase):
             with open_compressed(compressed_file) as f:
                 compressed_lines = [line.strip() for line in f]
             assert compressed_lines == uncompressed_lines
+
+
+class TestCacheFile(AllenNlpTestCase):
+    def test_temp_file_removed_on_error(self):
+        cache_filename = self.TEST_DIR / "cache_file"
+        with pytest.raises(IOError, match="I made this up"):
+            with CacheFile(cache_filename) as handle:
+                raise IOError("I made this up")
+        assert not os.path.exists(handle.name)
+        assert not os.path.exists(cache_filename)