+++ CHANGELOG.md
@@ -9,6 +9,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Reduced the amount of log messages produced by `allennlp.common.file_utils`.
+
+## [v1.0.0](https://github.com/allenai/allennlp/releases/tag/v1.0.0) - 2020-06-16
+
+### Fixed
+
+- Lazy dataset readers now work correctly with multi-process data loading.
+- Fixed race conditions that could occur when using a dataset cache.
+
+### Added
+
+- A bug where where all datasets would be loaded for vocab creation even if not needed.
+- A parameter to the `DatasetReader` class: `manual_multi_process_sharding`. This is similar
+  to the `manual_distributed_sharding` parameter, but applies when using a multi-process
+  `DataLoader`.
+
+## [v1.0.0rc6](https://github.com/allenai/allennlp/releases/tag/v1.0.0rc6) - 2020-06-11
+
+### Fixed
+
 - A bug where `TextField`s could not be duplicated since some tokenizers cannot be deep-copied.
   See https://github.com/allenai/allennlp/issues/4270.
 - Our caching mechanism had the potential to introduce race conditions if multiple processes
@@ -23,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Predictor.from_path` now automatically loads plugins (unless you specify `load_plugins=False`) so
   that you don't have to manually import a bunch of modules when instantiating predictors from
   an archive path.
+- `allennlp-server` automatically found as a plugin once again.
 
 ### Added